geoApp/requirement.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
Django==4.1.2
docopt==0.6.2
Fiona @ file:///C:/Users/Sushant/pipwin/Fiona-1.8.21-cp39-cp39-win_amd64.whl
fonttools==4.38.0
GDAL @ file:///C:/Users/Sushant/pipwin/GDAL-3.4.3-cp39-cp39-win_amd64.whl
GeoAlchemy2==0.12.5
geopandas==0.12.0
geoserver-rest==2.3.4
greenlet==1.1.3.post0
idna==3.4
Js2Py==0.71
kiwisolver==1.4.4
matplotlib==3.6.0
munch==2.5.0
numpy @ file:///C:/Users/Sushant/pipwin/numpy-1.22.4%2Bvanilla-cp39-cp39-win_amd64.whl
packaging==21.3
pandas @ file:///C:/Users/Sushant/pipwin/pandas-1.4.3-cp39-cp39-win_amd64.whl
Pillow==9.2.0
pipwin==0.5.2
psycopg2==2.9.5
Pygments==2.13.0
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pyproj @ file:///C:/Users/Sushant/pipwin/pyproj-3.3.1-cp39-cp39-win_amd64.whl
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
Rtree @ file:///C:/Users/Sushant/pipwin/Rtree-1.0.0-cp39-cp39-win_amd64.whl
seaborn==0.12.1
Shapely @ file:///C:/Users/Sushant/pipwin/Shapely-1.8.2-cp39-cp39-win_amd64.whl
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
sqlparse==0.4.3
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
xmltodict==0.13.0
